start structuring the sidebar
